fix: keep sample deck workflow reusable
