+++ src/features/sample-deck/index.ts
@@ -1,0 +1,1 @@
+export { prepareSampleDeck, useAddSampleDeck } from "./model/useAddSampleDeck";
